:rocket: Assembler: Start tinkering with code generation
There is still much to learn and to improve. We can split way better the
little components that make up the system. Create small, but readable,
functions to improve the readability of the code, especially in
`compile_function_body`. Now it should be able to compile the basic
program you can find in examples/basic.rack